mAIN command

+++ src/main/java/me/magandgolden/MainCommand.java
@@ -80,18 +80,46 @@ public class MainCommand extends BaseCommand {
 		String uuid = (sender instanceof Player) ? ((Player) sender).getUniqueId().toString() : "console";
 		switch (action) {
 			case "send":
-				dw.sendRandom(sender);
+				dw.sendRandom(sender); // Does not need to pass list or file bc it was passed to dw above.
 				break;
 			case "add":
-				dw.addFile(sender, uuid, numberOrText);
+				if (! sender.hasPermission("quotes.add")) {  // Check if sender has permission
+					noPermission(sender, "quotes.add");  // if sender does NOT have permission
+					return;// Pass sender and the permission then stop.
+				}
+				if (numberOrText == null) {
+					// If you don't check and it passes null will throw exception
+					sender.sendMessage(ChatColor.RED + "You need to enter something to add");
+					return;
+				}
+				if (! dw.userOverLimit(uuid, "quotes")) // Check how many jokes a user has added
+					dw.addFile(sender, uuid, numberOrText);
+				else {
+					sender.sendMessage(ChatColor.translateAlternateColorCodes('&', // New line is ignored
+							"&fPlease remove a joke before adding more")); // Java continues reading until ;
+					return;
+				}
+				break;
 			case "list":
+				if (! sender.hasPermission("quotes.list")) {
+					noPermission(sender, "quotes.list");
+					return;
+				}
 				dw.sendList(sender);
 				break;
 			case "remove":
+				if (! sender.hasPermission("quotes.remove")) {
+					noPermission(sender, "quotes.remove");
+					return;
+					// If you don't check and it passes null will throw exception
+				if (numberOrText == null) {
+						sender.sendMessage(ChatColor.RED + "You need to specify what to remove:");
+						dw.sendList(sender);
+						return;
+					}
 				int number = Integer.parseInt(numberOrText);
-				dw.removeFile(sender, number);
-				dw.sendList(sender);
-		}
+					dw.removeFile(sender, number);
+				}
 	}
 
 	private static void noPermission (CommandSender sender, String permission) {
